PAC-47 Pause menu

+++ src/main/java/org/example/app/components/pauseMenu/PauseMenuPanel.java
@@ -22,40 +22,40 @@ public class PauseMenuPanel extends JPanel {
         g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
         g2.setColor(Color.RED);
         g2.setFont(new Font("Impact", Font.BOLD, 100));
-        g2.drawString("Pure & Corrupt",(int) (MapConstants.GRID_CELL_SIZE * 3), (int) (MapConstants.GRID_CELL_SIZE * 3));
+        g2.drawString("Pure & Corrupt", (MapConstants.GRID_CELL_SIZE * 3), (MapConstants.GRID_CELL_SIZE * 3));
         g2.setFont(new Font("Impact", Font.BOLD, 200));
-        g2.drawString("PAUSED",(int) (MapConstants.GRID_CELL_SIZE * 3), (int) (MapConstants.GRID_CELL_SIZE * 8));
+        g2.drawString("PAUSED", (MapConstants.GRID_CELL_SIZE * 3), (MapConstants.GRID_CELL_SIZE * 8));
 
         g2.setFont(new Font("Impact", Font.PLAIN, 48));
-        g2.drawString("Controls:",(int) (MapConstants.GRID_CELL_SIZE * 3), (int) (MapConstants.GRID_CELL_SIZE * 10));
+        g2.drawString("Controls:", (MapConstants.GRID_CELL_SIZE * 3), (MapConstants.GRID_CELL_SIZE * 10));
         g2.setColor(Color.WHITE);
         g2.setFont(new Font("Impact", Font.PLAIN, 32));
-        g2.drawString("WASD - move",(int) (MapConstants.GRID_CELL_SIZE * 3), (int) (MapConstants.GRID_CELL_SIZE * 11));
-        g2.drawString("SHIFT - attack",(int) (MapConstants.GRID_CELL_SIZE * 3), (int) (MapConstants.GRID_CELL_SIZE * 12));
-        g2.drawString("Q, E - activate items",(int) (MapConstants.GRID_CELL_SIZE * 3), (int) (MapConstants.GRID_CELL_SIZE * 13));
-        g2.drawString("ESC - pause",(int) (MapConstants.GRID_CELL_SIZE * 3), (int) (MapConstants.GRID_CELL_SIZE * 14));
+        g2.drawString("WASD - move", (MapConstants.GRID_CELL_SIZE * 3), (MapConstants.GRID_CELL_SIZE * 11));
+        g2.drawString("SHIFT - attack", (MapConstants.GRID_CELL_SIZE * 3), (MapConstants.GRID_CELL_SIZE * 12));
+        g2.drawString("Q, E - activate items", (MapConstants.GRID_CELL_SIZE * 3), (MapConstants.GRID_CELL_SIZE * 13));
+        g2.drawString("ESC - pause", (MapConstants.GRID_CELL_SIZE * 3), (MapConstants.GRID_CELL_SIZE * 14));
 
         g2.setColor(Color.RED);
-        double LORE_OFFSET = 12;
+        int LORE_OFFSET = 12;
         g2.setFont(new Font("Impact", Font.PLAIN, 48));
-        g2.drawString("Lore:",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 10));
+        g2.drawString("Lore:", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 10));
         g2.setColor(Color.WHITE);
         g2.setFont(new Font("Impact", Font.PLAIN, 16));
-        g2.drawString("Hell is real. And so is Heaven, but people seem like they've forgotten that. The world was taken over by sin.",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 11));
-        g2.drawString("That fueled the forces of Hell to the point where they became much stronger than Heaven and humans ever could imagine.",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 12));
-        g2.drawString("Hell's corruption became so powerful, that the very ground beneath places important to the Church trembled until it ",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 13));
-        g2.drawString("broke and sank underground, closer to Hell itself. Demons, Chaos spawn, Corrupted folk, all invaded the lands of humans ",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 14));
-        g2.drawString("and spread fear and destruction. Humans fled right back to what they abandoned.",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 15));
-        g2.drawString("The Highest Church, worshiping the merciful Heaven. Church was the only thing standing between them and total destruction. ",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 16));
-        g2.drawString("The Church and their soldiers. Paladins. Such as you. The best of the best. The most pure and uncorrupted of them all.",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 17));
-        g2.drawString("And the Church chose you specifically to go down where it all started. Through the sunken ground, right to the center of Hell and end it all for good.",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 18));
-        g2.drawString("What they forgot to tell you are the bodies of your comrades that have fallen on the very same mission.\n",(int) (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (int) (MapConstants.GRID_CELL_SIZE * 19));
+        g2.drawString("Hell is real. And so is Heaven, but people seem like they've forgotten that. The world was taken over by sin.", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 11));
+        g2.drawString("That fueled the forces of Hell to the point where they became much stronger than Heaven and humans ever could imagine.", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 12));
+        g2.drawString("Hell's corruption became so powerful, that the very ground beneath places important to the Church trembled until it ", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 13));
+        g2.drawString("broke and sank underground, closer to Hell itself. Demons, Chaos spawn, Corrupted folk, all invaded the lands of humans ", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 14));
+        g2.drawString("and spread fear and destruction. Humans fled right back to what they abandoned.", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 15));
+        g2.drawString("The Highest Church, worshiping the merciful Heaven. Church was the only thing standing between them and total destruction. ", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 16));
+        g2.drawString("The Church and their soldiers. Paladins. Such as you. The best of the best. The most pure and uncorrupted of them all.", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 17));
+        g2.drawString("And the Church chose you specifically to go down where it all started. Through the sunken ground, right to the center of Hell and end it all for good.", (MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 18));
+        g2.drawString("What they forgot to tell you are the bodies of your comrades that have fallen on the very same mission.\n",(MapConstants.GRID_CELL_SIZE * LORE_OFFSET), (MapConstants.GRID_CELL_SIZE * 19));
 
         g2.setColor(Color.RED);
-        g2.drawRect((MapConstants.GRID_CELL_SIZE * 30), (int) (MapConstants.GRID_CELL_SIZE),(MapConstants.GRID_CELL_SIZE * 9), (int) (MapConstants.GRID_CELL_SIZE * 2));
+        g2.drawRect((MapConstants.GRID_CELL_SIZE * 30), (MapConstants.GRID_CELL_SIZE),(MapConstants.GRID_CELL_SIZE * 9),(MapConstants.GRID_CELL_SIZE * 2));
         g2.setFont(new Font("Impact", Font.BOLD, 48));
         g2.drawString("QUIT GAME",(int) (MapConstants.GRID_CELL_SIZE * 31.5), (int) (MapConstants.GRID_CELL_SIZE * 2.5));
-        g2.drawRect((MapConstants.GRID_CELL_SIZE * 30), (int) (MapConstants.GRID_CELL_SIZE * 4),(MapConstants.GRID_CELL_SIZE * 9), (int) (MapConstants.GRID_CELL_SIZE * 2));
+        g2.drawRect((MapConstants.GRID_CELL_SIZE * 30), (MapConstants.GRID_CELL_SIZE * 4),(MapConstants.GRID_CELL_SIZE * 9), (MapConstants.GRID_CELL_SIZE * 2));
         g2.drawString("RESUME",(int) (MapConstants.GRID_CELL_SIZE * 32.25), (int) (MapConstants.GRID_CELL_SIZE * 5.5));
 
     }
